DAO layer, BUS layer, Shift GUI, Roboto Custom Font

+++ DTO/ChiTietKhuyenMaiDTO.java
@@ -3,11 +3,11 @@ package DTO;
 public class ChiTietKhuyenMaiDTO {
     private String maKhuyenMai;
     private String maSanPham;
-    private Double giaTriKhuyenMai;
+    private Float giaTriKhuyenMai;
 
 
     public ChiTietKhuyenMaiDTO() {}
-    public ChiTietKhuyenMaiDTO(String maKhuyenMai, String maHoaDon, Double  giaTriKhuyenMai) {
+    public ChiTietKhuyenMaiDTO(String maKhuyenMai, String maHoaDon, Float giaTriKhuyenMai) {
         this.maKhuyenMai = maKhuyenMai;
         this.maSanPham = maHoaDon;
         this.giaTriKhuyenMai = giaTriKhuyenMai;
@@ -29,11 +29,11 @@ public class ChiTietKhuyenMaiDTO {
         this.maSanPham = maSanPham;
     }
 
-    public Double getGiaTriKhuyenMai() {
+    public Float getGiaTriKhuyenMai() {
         return giaTriKhuyenMai;
     }
 
-    public void setGiaTriKhuyenMai(Double giaTriKhuyenMai) {
+    public void setGiaTriKhuyenMai(Float giaTriKhuyenMai) {
         this.giaTriKhuyenMai = giaTriKhuyenMai;
     }
 }
